+++ src/main/java/com/example/week06/domain/user/dto/UserInfoResponseDto.java
@@ -1,7 +1,7 @@
-package com.example.week06.dto;
+package com.example.week06.domain.user.dto;
 
-import com.example.week06.model.User;
-import com.example.week06.security.UserDetailsImpl;
+import com.example.week06.domain.user.entity.User;
+import com.example.week06.global.security.UserDetailsImpl;
 import lombok.*;
