+++ src/main/java/com/example/week06/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.week06.repository;
+package com.example.week06.domain.user.repository;
 
-import com.example.week06.model.User;
+import com.example.week06.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
